Split MethylVI/MethylANVI models/modules into separate files
